+++ src/daft-core/src/array/mod.rs
@@ -20,7 +20,6 @@ pub use fixed_size_list_array::FixedSizeListArray;
 pub use list_array::ListArray;
 pub use struct_array::StructArray;
 mod boolean;
-mod from_iter;
 pub mod prelude;
 use std::{marker::PhantomData, sync::Arc};
