+++ src/daft-functions-uri/Cargo.toml
@@ -1,4 +1,5 @@
 [dependencies]
+arrow = {workspace = true}
 async-trait = {workspace = true}
 bytes = {workspace = true}
 common-error = {path = "../common/error", default-features = false}
