+++ src/daft-parquet/Cargo.toml
@@ -1,4 +1,5 @@
 [dependencies]
+arrow = {workspace = true}
 async-compat = {workspace = true}
 daft-arrow = {path = "../daft-arrow"}
 async-stream = {workspace = true}
